+++ 05-arrays_slices_maps/01-arrays/02-float/02/main.go
@@ -5,9 +5,9 @@ import (
 )
 
 func main() {
-	x := [6]float32{98.1, 93.54, 7.21, 2.76, 1.2}
+	x := [6]float64{98.1, 93.54, 7.21, 2.76, 1.2}
 
 	fmt.Println(x)
-	fmt.Println("Lenght of the array:", len(x)) // We print the lenght of the array
+	fmt.Println("Lenght of the array:", len(x))
 	fmt.Printf("Type-%T\n", x)
 }
